backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
pydantic==2.5.0
pydantic-settings==2.1.0
openai==1.6.1
Pillow==10.4.0
numpy==1.26.4
sqlalchemy==2.0.23
aiosqlite==0.19.0
boto3==1.34.0
google-cloud-storage==2.10.0